+++ src/hooks/transactions/useSignTransactionsWithDevice.tsx
@@ -17,6 +17,7 @@ import {
   DeviceSignedTransactions,
   LoginMethodsEnum,
   MultiSignTransactionType,
+  Nullable,
   TransactionBatchStatusesEnum
 } from 'types';
 import { getIsProviderEqualTo } from 'utils/account/getIsProviderEqualTo';
@@ -144,11 +145,15 @@ export function useSignTransactionsWithDevice(
     clearTransactionsToSignWithWarning(sessionId);
   }
 
-  async function handleSignTransaction(transaction: Transaction) {
+  async function handleSignTransaction(transaction: Nullable<Transaction>) {
     const connectedProvider =
       providerType !== LoginMethodsEnum.ledger
         ? provider
         : await getLedgerProvider();
+
+    if (!transaction) {
+      return null;
+    }
 
     return await connectedProvider.signTransaction(transaction);
   }