+++ src/types/dappProvider.types.ts
@@ -1,5 +1,6 @@
 import { SignableMessage, Transaction } from '@multiversx/sdk-core';
 import { EngineTypes } from 'utils/walletconnect/__sdkWalletconnectProvider';
+import { Nullable } from './transactions.types';
 
 export interface DappOptions {
   callbackUrl?: string;
@@ -7,7 +8,7 @@ export interface DappOptions {
 
 export interface IDappProvider {
   init?(): Promise<boolean>;
-  login?(options?: DappOptions): Promise<string>;
+  login?(options?: DappOptions): Promise<string | boolean>;
   logout(options?: DappOptions): Promise<boolean>;
   getAddress(): Promise<string>;
   isInitialized(): boolean;
@@ -19,15 +20,15 @@ export interface IDappProvider {
   signTransaction(
     transaction: Transaction,
     options?: DappOptions
-  ): Promise<Transaction>;
+  ): Promise<Nullable<Transaction | undefined>>;
   signTransactions(
     transactions: Transaction[],
     options?: DappOptions
-  ): Promise<Transaction[]>;
-  signMessage<T extends SignableMessage>(
-    message: T,
+  ): Promise<Nullable<Transaction[]>>;
+  signMessage(
+    message: SignableMessage,
     options: DappOptions
-  ): Promise<T>;
+  ): Promise<Nullable<SignableMessage>>;
   sendCustomMessage?({
     method,
     params
@@ -39,5 +40,5 @@ export interface IDappProvider {
     request: EngineTypes.RequestParams['request'];
   }): Promise<any>;
   ping?(): Promise<boolean>;
-  relogin?: () => Promise<void>;
+  relogin?: () => Promise<Nullable<string | void>>;
 }